Try to fix issue with type mismatch on numpy 1.X + windows

+++ pynbody/chunk/__init__.py
@@ -85,7 +85,7 @@ class LoadControl:
 
         # generate simulation-level ID list
         if hasattr(clauses, "__len__"):
-            self._ids = np.asarray(clauses)
+            self._ids = np.asarray(clauses, dtype=np.int64) 
         else:
             self._ids = None  # no partial loading!
